refactor(overview): extract shared useStats hook for legacy + v2 headers
Both headers had their own copy of the bucket size/object/package stats hook
(v2 carried a 'keep in sync' NOTE). Move it to Overview/useStats.ts returning
the superset both consume; legacy destructures the subset it needs.
